requirements.txt
streamlit
requests
# NLP
nltk==3.8
pdfplumber
python-docx
matplotlib
seaborn
wordcloud
boto3
